pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mihomo-update"
version = "0.5.1"
description = "Update mihomo proxies from subscription"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[project.scripts]
mihomo-update = "mihomo_update.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mihomo_update = ["locale/**/*.mo"]
